Move fetcher to global place
This is used in multiple places
